feat: 組件化目前畫面
